requirements.txt
# Core web framework
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.28.0

# AI/ML Libraries
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.1.0
elevenlabs>=1.0.0

# Web scraping and HTTP
requests>=2.28.0
beautifulsoup4>=4.11.0
httpx>=0.24.0

# Audio processing
gtts>=2.3.0

# Async and rate limiting
aiolimiter>=1.1.0
tenacity>=8.0.0

# Data handling
pydantic>=2.0.0

# Environment and utilities
python-dotenv>=1.0.0

# MCP (Model Context Protocol)
mcp>=1.0.0